Rename a component for clarity
