+++ src/popup/components/Popups/GroupMenuPopup.vue
@@ -7,7 +7,7 @@ import { useGroupStore } from '@/stores/group'
 import Popup from '@/components/Popups/Popup.vue'
 import AddLinkMenuItem from '@/components/Views/GroupView/GroupControls/MenuItems/AddLinkMenuItem.vue'
 import DeleteGroupMenuItem from '@/components/Views/GroupView/GroupControls/MenuItems/DeleteGroupMenuItem.vue'
-import RenameMenuItem from '@/components/Views/GroupView/GroupControls/MenuItems/RenameMenuItem.vue'
+import RenameGroupMenuItem from '@/components/Views/GroupView/GroupControls/MenuItems/RenameGroupMenuItem.vue'
 import RebindUrlItem from '@/components/Views/GroupView/GroupControls/MenuItems/RebindUrlItem.vue'
 import ExportGroupMenuItem from '@/components/Views/GroupView/GroupControls/MenuItems/ExportGroupMenuItem.vue'
 import MenuItem from '@/components/MenuItem.vue'
@@ -35,7 +35,7 @@ const isEncrypted = computed<boolean>(() => {
 
         <div v-else class="flex flex-col gap-1 mt-3">
             <AddLinkMenuItem />
-            <RenameMenuItem />
+            <RenameGroupMenuItem />
             <ExportGroupMenuItem />
 
             <RouterLink
