+++ src/popup/components/Views/GroupView/GroupControls/MenuItems/RenameGroupMenuItem.vue
